requirements.txt
absl-py==0.9.0
appnope==0.1.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kaggle==1.5.6
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
netron
notebook==6.0.3
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-slugify==4.0.0
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
sentry-sdk==0.14.4
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
subprocess32==3.5.4
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.24.3
wandb==0.8.36
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
